pyproject.toml
[project]
name = "engageai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.4.0",
    "fastapi>=0.115.11",
    "openai>=1.67.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "pytz>=2025.1",
    "redis>=5.2.1",
    "sqlalchemy>=2.0.39",
    "tenacity>=9.1.2",
    "twilio>=9.5.0",
    "uvicorn>=0.34.0",
]
